+++ examples/lambda-url/README.md
@@ -2,10 +2,22 @@
 
 This example must be deployed to AWS to work.
 
-`$ yarn install`
-`$ yarn build`
-`$ yarn deploy`
+```sh
+pnpm install
+pnpm build
+pnpm deploy
+```
 
-## Run the client
+This will deploy the Lambda function and create a Lambda URL in the format of `https://<your-lambda-url>.lambda-url.us-east-1.on.aws/`.
 
-`$ yarn start`
+## Server
+
+The server is a simple TRPC server that uses the `awslambda.streamifyResponse()` function to stream the response to the client. See this [blog post](https://aws.amazon.com/blogs/compute/introducing-aws-lambda-response-streaming/) for more information about AWS Lambda response streaming.
+
+## Client
+
+Replace `YOUR_LAMBDA_URL` in `src/client.ts` with the Lambda URL you just created.
+
+```sh
+pnpm start
+```
