+++ packages/server/src/adapters/aws-lambda/getPlanner.ts
@@ -5,8 +5,31 @@ import type {
   APIGatewayProxyEventV2,
   APIGatewayProxyResult,
   APIGatewayProxyStructuredResultV2,
+  Context,
 } from 'aws-lambda';
 import { splitSetCookieString } from '../../vendor/cookie-es/set-cookie/split';
+
+declare global {
+  namespace awslambda {
+    class HttpResponseStream extends Writable {
+      static from(
+        writable: Writable,
+        metadata: Record<string, unknown>,
+      ): HttpResponseStream;
+      setContentType: (contentType: string) => void;
+    }
+
+    type StreamifyHandler<TEvent = any, TResult = any> = (
+      event: TEvent,
+      responseStream: awslambda.HttpResponseStream,
+      context: Context,
+    ) => TResult | Promise<TResult>;
+
+    function streamifyResponse<TEvent = any, TResult = void>(
+      handler: StreamifyHandler<TEvent, TResult>,
+    ): StreamifyHandler<TEvent, TResult>;
+  }
+}
 
 export type LambdaEvent = APIGatewayProxyEvent | APIGatewayProxyEventV2;
 
@@ -32,8 +55,8 @@ function determinePayloadFormat(event: LambdaEvent): string {
 export type inferAPIGWReturn<TEvent> = TEvent extends APIGatewayProxyEvent
   ? APIGatewayProxyResult
   : TEvent extends APIGatewayProxyEventV2
-    ? APIGatewayProxyStructuredResultV2
-    : never;
+  ? APIGatewayProxyStructuredResultV2
+  : never;
 
 interface Processor<TEvent extends LambdaEvent> {
   getTRPCPath: (event: TEvent) => string;
